Initial complete closed loop boost control. NOT TESTED!
Only works with kpa in TunerStudio.

+++ libs/PID_v1/PID_v1.h
@@ -90,7 +90,7 @@ class integerPID
   #define REVERSE  1
 
   //commonly used functions **************************************************************************
-    integerPID(long*, long*, long*,        // * constructor.  links the PID to the Input, Output, and 
+    integerPID(unsigned int*, long*, long*,        // * constructor.  links the PID to the Input, Output, and 
         byte, byte, byte, byte);     //   Setpoint.  Initial tuning parameters are also set here
 	
     void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
